requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
fastmcp
faiss-cpu
langchain_openai
sentence-transformers